% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arenosa_ld_example}
\alias{arenosa_ld_example}
\title{Example LD Output for Autotetraploid Arabidopsis arenosa}
\format{
A data frame containing estimation results.
}
\source{
Simulated/Empirical subset
}
\usage{
arenosa_ld_example
}
\description{
A subset of linkage disequilibrium estimation results generated by AutoLD.
}
\keyword{datasets}
