% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{auto2_est}
\alias{auto2_est}
\title{Estimate Linkage Disequilibrium for Diploids}
\usage{
auto2_est(dose)
}
\arguments{
\item{dose}{A numeric vector or factor indicating the 9 possible joint genotype classes for two loci.}
}
\value{
A list containing:
\itemize{
\item \code{p}: Estimated allele frequency for locus A.
\item \code{q}: Estimated allele frequency for locus B.
\item \code{D}: Estimated Linkage Disequilibrium parameter.
\item \code{iter}: Number of EM iterations until convergence.
}
}
\description{
This function estimates linkage disequilibrium (LD) parameter D and allele frequencies
for diploids using an Expectation-Maximization (EM) algorithm.
}
\examples{
\dontrun{
true_para <- c(0.6, 0.4, 0.05) # p, q, D
sim_data <- sim_2locus_2x_cLD(true_para, n=1000)
est_res <- auto2_est(sim_data)
print(est_res)
}
}
