% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_statistics.R
\name{LD_test_optimized_final}
\alias{LD_test_optimized_final}
\title{Highly Optimized Likelihood Ratio Test for LD (Accounting for Double Reduction)}
\usage{
LD_test_optimized_final(
  res,
  geno,
  ploidy = 4,
  type = "full",
  hwdA = 0,
  hwdB = 0
)
}
\arguments{
\item{res}{A numeric vector of parameter estimates under the alternative hypothesis (H1).}

\item{geno}{A numeric vector or data frame column of joint genotype classes.}

\item{ploidy}{An integer specifying the ploidy level (e.g., 4 or 6). Default is 4.}

\item{type}{A character string specifying the resolution model: \code{"full"} or \code{"partial"}. Default is \code{"full"}.}

\item{hwdA}{Numeric. Hardy-Weinberg Disequilibrium parameter for locus A. Default is 0.}

\item{hwdB}{Numeric. Hardy-Weinberg Disequilibrium parameter for locus B. Default is 0.}
}
\value{
A named numeric vector of p-values corresponding to the tested LD parameters.
}
\description{
The final, most highly optimized implementation of the Likelihood Ratio Test (LRT)
to evaluate the statistical significance of LD in autopolyploids. This function
natively accounts for double reduction rates estimated from the data.
}
