% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_6x_unified_hap}
\alias{sim_6x_unified_hap}
\title{Simulate Unified Haplotype Data for Autohexaploids}
\usage{
sim_6x_unified_hap(pall, n = 1000, mode = "partial")
}
\arguments{
\item{pall}{A numeric vector of length 3 containing the true parameters:
\code{pA} (allele frequency of A), \code{pB} (allele frequency of B), and \code{D} (overall LD).}

\item{n}{An integer specifying the sample size. Default is 1000.}

\item{mode}{A character string specifying the output format: \code{"full"} (49 distinct classes)
or \code{"partial"} (collapsed into 9 classes). Default is \code{"partial"}.}
}
\value{
A numeric vector representing the simulated joint genotype classes.
}
\description{
Generates simulated joint genotype data for autohexaploids using a unified haplotype-based
probability model. It handles the high dimensionality of hexaploid genetics by generating
data in either 49-class or 9-class resolutions.
}
\examples{
\dontrun{
hex_para <- c(0.6, 0.4, 0.05) # pA, pB, D
hex_data <- sim_6x_unified_hap(pall = hex_para, n = 1000, mode = "full")
table(hex_data)
}
}
